[BPM-178] 스튜디오 dto변환작업 dto에서 하지 않고 서비스레이어에서 하도록 변경
- 스튜디오 키워드쪽 변경, 리뷰쪽은 추후 작업
- 스튜디오 키워드로 필터링 가능하도록 API작

+++ src/main/java/d83t/bpmbackend/domain/aggregate/studio/dto/StudioFilterDto.java
@@ -10,5 +10,5 @@ import java.util.List;
 @Setter
 @ToString
 public class StudioFilterDto {
-    List<Integer> keyword;
+    List<Long> keyword;
 }
